Allow `local_assigns` variable w/ same-name as partial
Not sure why this is in there in the first place —  @alexch might be
able to shed some light on it if he’s around. (I’ve already looked at
the git blame, no help there.) I’m assuming it’s an ERB thing — it
seems that when rendering a partial in ERB, it adds an entry to the
local_assigns hash: `{ ‘[partial name]’  => nil }`

Not sure why it does this — but to preserve compatibility w/ ERB, we’ll
still ignore this if it’s nil.

closes #23

/cc @jcutrell to test this out

p.s. fuck ERB